+++ qbraid/transpiler/conversions/pytket/__init__.py
@@ -25,12 +25,19 @@ Functions
 
    pytket_to_braket
    pytket_to_cirq
+   pytket_to_pyqir
    pytket_to_qasm2
    pytket_to_qiskit
 
 """
 
-from .pytket_extras import pytket_to_braket, pytket_to_cirq, pytket_to_qiskit
+from .pytket_extras import pytket_to_braket, pytket_to_cirq, pytket_to_pyqir, pytket_to_qiskit
 from .pytket_to_qasm2 import pytket_to_qasm2
 
-__all__ = ["pytket_to_braket", "pytket_to_cirq", "pytket_to_qasm2", "pytket_to_qiskit"]
+__all__ = [
+    "pytket_to_braket",
+    "pytket_to_cirq",
+    "pytket_to_pyqir",
+    "pytket_to_qasm2",
+    "pytket_to_qiskit",
+]