+++ CHANGELOG.md
@@ -18,6 +18,8 @@ Types of changes:
 ### Added
 - Added `cirq_to_pytket` and `pytket_to_cirq` transpiler conversions (via the `pytket-cirq` extension), giving a direct Cirq <-> PyTKET edge in the conversion graph. Part of [#1200](https://github.com/qBraid/qBraid/issues/1200). ([#1208](https://github.com/qBraid/qBraid/pull/1208))
 
+- Added `pytket_to_pyqir` transpiler conversion (via the `pytket-qir` extension), giving a direct PyTKET -> PyQIR edge in the conversion graph. Part of [#1200](https://github.com/qBraid/qBraid/issues/1200). ([#1208](https://github.com/qBraid/qBraid/pull/1208))
+
 - Added `include_retired` parameter to `QbraidProvider.get_devices` method to optionally include retired devices in the device list ([#1201](https://github.com/qBraid/qBraid/pull/1201))
 
 - Added `PasqalProvider`, `PasqalDevice`, and `PasqalJob` classes implementing the qBraid runtime interface for Pasqal Cloud Services (neutral-atom QPUs and emulators, using Pulser as the native IR). Closes [#1185](https://github.com/qBraid/qBraid/issues/1185).